requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aioquic==0.9.25
aioquic-mitmproxy==0.9.21.1
aiosignal==1.3.2
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
blis==1.3.0
Brotli==1.1.0
bs4==0.0.1
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==42.0.5
cssselect==1.2.0
cymem==2.0.11
datasets==3.6.0
dill==0.3.8
fake-useragent==1.4.0
filelock==3.18.0
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.6.2
fsspec==2025.3.0
h11==0.14.0
h2==4.1.0
hf-xet==1.1.3
hpack==4.0.0
huggingface-hub==0.32.4
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.5.1
kaitaistruct==0.10
langcodes==3.5.0
language_data==1.3.0
ldap3==2.9.1
lxml==4.9.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mitmproxy==10.2.4
mitmproxy-windows==0.5.1
mitmproxy_rs==0.5.1
mpmath==1.3.0
msgpack==1.0.7
multidict==6.4.4
multiprocess==0.70.16
murmurhash==1.0.13
networkx==3.5
numpy==1.26.4
packaging==25.0
pandas==2.1.4
parse==1.20.0
passlib==1.7.4
pillow==10.2.0
preshed==3.0.10
propcache==0.3.1
protobuf==4.25.1
psutil==7.0.0
publicsuffix2==2.20191221
pyarrow==20.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.11.5
pydantic_core==2.33.2
pydivert==2.1.0
pyee==8.2.2
Pygments==2.19.1
pylsqpack==0.3.18
PyMySQL==1.1.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyperclip==1.8.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.1.0
requests==2.32.3
requests-html==0.10.0
rich==14.0.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
service-identity==23.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.1
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.4
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2023.3
urllib3==1.26.18
urwid-mitmproxy==2.1.2.1
w3lib==2.1.2
wasabi==1.1.3
weasel==0.4.1
websockets==10.4
Werkzeug==3.0.1
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.17.0
zstandard==0.22.0
